Cascade search filter includes
* rename property_filterX files into search_filterX in coreobjects
